extractor-function/requirements.txt
functions-framework==3.*
requests
google-cloud-documentai
google-cloud-firestore
